alpha

+++ src/app/items/item.component.ts
@@ -1,7 +1,7 @@
 import {Component} from '@angular/core';
 
 import {ItemService} from './item.service';
-
+import {Router} from '@angular/router-deprecated';
 // We `import` `http` into our `ItemService` but we can only
 // specify providers within our component
 import {HTTP_PROVIDERS} from '@angular/http';
@@ -17,6 +17,7 @@ import {NgFor} from '@angular/common';
     providers: [...HTTP_PROVIDERS, ItemService],
     template: require('./item.html')
 })
+
 export class Item {
 
   // Initialize our `itemData.text` to an empty `string`
@@ -31,27 +32,61 @@ export class Item {
 
   private items: Array<Item> = [];
   private rukus: Array<any> = [];
+  private chukus: Array<any> = [];
   private arrays: Array<number> = [];
-  constructor(public itemService: ItemService) {
+  private clicked: boolean = false;
+  constructor(private router: Router, public itemService: ItemService) {
     console.log('Item constructor go!');
+      if (localStorage.getItem('token')) {
+          console.log(JSON.parse(localStorage.getItem('token')));
+      } else {
 
-      itemService.getAll()
-        // `Rxjs`; we subscribe to the response
-        .subscribe((res) => {
+        router.navigate(['Index']);
+      }
+                itemService.getAll()
+                  // `Rxjs`; we subscribe to the response
+                  .subscribe((res) => {
 
-            // Populate our `item` array with the `response` data
-            this.items = res;
-            // Reset `item` input
-            this.itemData.text = '';
-            this.itemData.name = '';
-            this.itemData.price = null;
-            this.itemData.number = 0;
-            this.itemData.saleprice = null;
-        });
+                      // Populate our `item` array with the `response` data
+                      this.items = res;
+                      // Reset `item` input
+                      this.itemData.text = '';
+                      this.itemData.name = '';
+                      this.itemData.price = null;
+                      this.itemData.number = 0;
+                      this.itemData.saleprice = null;
+                  });
+       let t = setInterval(() => {
+
+          if(window.location.hash === "#/item"){
+
+            if (this.clicked === true){
+                itemService.getAll()
+                  // `Rxjs`; we subscribe to the response
+                  .subscribe((res) => {
+
+                      // Populate our `item` array with the `response` data
+                      this.items = res;
+                      // Reset `item` input
+                      this.itemData.text = '';
+                      this.itemData.name = '';
+                      this.itemData.price = null;
+                      this.itemData.number = 0;
+                      this.itemData.saleprice = null;
+                  });
+                 this.clicked = false;
+            }
+          }else{
+
+              clearInterval(t);
+          }
+
+      }, 1000);
   }
 
   createItem() {
-
+      if (JSON.parse(localStorage.getItem('token')).role === "admin") {
+        console.log("111");
       this.itemService.createItem(this.itemData)
         .subscribe((res) => {
 
@@ -64,18 +99,22 @@ export class Item {
             this.itemData.number = 0;
             this.itemData.saleprice = null;
         });
+      }
   }
 
   deleteItem(id) {
+      if (JSON.parse(localStorage.getItem('token')).role === "admin") {
 
-    this.itemService.deleteItem(id)
-      .subscribe((res) => {
+        this.itemService.deleteItem(id)
+          .subscribe((res) => {
 
-          // Populate our `item` array with the `response` data
-          this.items = res;
-      });
+              // Populate our `item` array with the `response` data
+              this.items = res;
+          });
+      }
   }
   updateItems(){
+
     this.itemService.getAllruku()
       .subscribe((res) => {
 
@@ -90,10 +129,33 @@ export class Item {
             this.arrays[ruku.name] = this.arrays[ruku.name] + ruku.number
         }
         console.log(this.arrays);
+              this.itemService.getAllchuku()
+                .subscribe((res) => {
 
+                  // Populate our `item` array with the `response` data
+                  this.chukus = res;
+                  console.log(this.chukus);
+                  for (let chuku of this.chukus){
+                      console.log(chuku.number);
+                      if(isNaN(this.arrays[chuku.name])){
+                        this.arrays[chuku.name] = 0;
+                      }
+                      this.arrays[chuku.name] = this.arrays[chuku.name] - chuku.number
+                  }
+                  console.log(this.arrays);
+                  for (let array in this.arrays){
+                      console.log(array);
+                      console.log(this.arrays[array]);
+                      this.itemService.updateItem(array,{number: this.arrays[array]})
+                      .subscribe((res) => {
 
-
+                      });
+                  }
+        this.clicked = true;
         this.arrays = [];
+
+      });
+
       });
 
   }
